Move HealthCheck to the httputil package
