+++ internal/setup/base.go
@@ -15,8 +15,6 @@
 package setup
 
 import (
-	"database/sql"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
@@ -25,7 +23,6 @@ import (
 
 	"github.com/matrix-org/dendrite/internal/caching"
 	"github.com/matrix-org/dendrite/internal/httputil"
-	"github.com/matrix-org/dendrite/internal/sqlutil"
 	"github.com/matrix-org/gomatrixserverlib"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 
@@ -270,7 +267,7 @@ func (b *BaseDendrite) CreateFederationClient() *gomatrixserverlib.FederationCli
 // nolint:gocyclo
 func (b *BaseDendrite) SetupAndServeHTTP(
 	internalHTTPAddr, externalHTTPAddr config.HTTPAddress,
-	certFile, keyFile *string,
+	certFile, keyFile *string, dbConfigs ...config.DatabaseOptions,
 ) {
 	internalAddr, _ := internalHTTPAddr.Address()
 	externalAddr, _ := externalHTTPAddr.Address()
@@ -298,6 +295,8 @@ func (b *BaseDendrite) SetupAndServeHTTP(
 	if b.Cfg.Global.Metrics.Enabled {
 		internalRouter.Handle("/metrics", httputil.WrapHandlerInBasicAuth(promhttp.Handler(), b.Cfg.Global.Metrics.BasicAuth))
 	}
+
+	b.InternalAPIMux.HandleFunc("/health", httputil.HealthCheckHandler(dbConfigs...))
 
 	externalRouter.PathPrefix(httputil.PublicClientPathPrefix).Handler(b.PublicClientAPIMux)
 	externalRouter.PathPrefix(httputil.PublicKeyPathPrefix).Handler(b.PublicKeyAPIMux)
@@ -363,61 +362,4 @@ func setupNaffka(cfg *config.Dendrite) (sarama.Consumer, sarama.SyncProducer) {
 		logrus.WithError(err).Panic("Failed to setup naffka")
 	}
 	return naff, naff
-}
-
-// healthResponse is returned on requests to /api/health
-type healthResponse struct {
-	Code       int    `json:"code"`
-	FirstError string `json:"error"`
-}
-
-// AddHealthCheck adds a /health endpoint to the internal api mux
-func AddHealthCheck(apiMux *mux.Router, dbConfig ...config.DatabaseOptions) {
-	if len(dbConfig) == 0 {
-		return
-	}
-	conns := make([]*sql.DB, len(dbConfig))
-	// populate sql connections
-	for i, conf := range dbConfig {
-		c, err := sqlutil.Open(&conf)
-		if err != nil {
-			panic(err)
-		}
-		conns[i] = c
-	}
-
-	apiMux.HandleFunc("/health", func(resp http.ResponseWriter, _ *http.Request) {
-		var (
-			errMsg string
-			code   = http.StatusOK
-		)
-		err := dbPingCheck(conns, &errMsg, &code)
-		if err != nil {
-			resp.WriteHeader(http.StatusInternalServerError)
-		}
-		data, err := json.Marshal(healthResponse{
-			Code:       code,
-			FirstError: errMsg,
-		})
-		if err != nil {
-			logrus.WithError(err).Error("Unable to encode response")
-			resp.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		if _, err = resp.Write(data); err != nil {
-			logrus.WithError(err).Error("Unable to write health response")
-		}
-	})
-}
-
-func dbPingCheck(conns []*sql.DB, errMsg *string, code *int) error {
-	// check every database connection
-	for _, conn := range conns {
-		if err := conn.Ping(); err != nil {
-			*errMsg = err.Error()
-			*code = http.StatusInternalServerError
-			return err
-		}
-	}
-	return nil
 }
